Update dependencies, move code
